The code had a completely duplicated set of constants for video modes and frame rates, which is then mapped to the DC1394 versions. There is no need for this so I deleted all the local constants and now we use the values in the dc1394 library. It is much easier to understand the code now. This also involved me deleting the int_* elements from the big camera config struct because they are not needed any more.
The second major change is that the default video mode was YUV411_HALF, and
by default the simpleTest application does not work. So I'm not sure when
this did work or if something else changed, but I have now modified it so
that the default mode is 640x480 with RGB. You can always override this with
a vconf string, but this will make all the normal applications work by default
now like it should.


git-svn-id: svn://svn.code.sf.net/p/artoolkit/code/trunk@260 d65a7696-613b-0410-ac10-aa9ec39ffaa2